Games with missing dev Uuid
How to help: Run `adb logcat` and go launch the game. May need to purchase something? Logcat will show the developer UUID. 

Go to the "classic" folder and find the corresponding packageName for the game and edit that file. Change the Devleoper UUID from 00000000-... to whatever was displayed by logcat.